# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login a user",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/campaigns": {
      "get": {
        "operationId": "CampaignsController_findAll",
        "parameters": [
          {
            "name": "activeOnly",
            "required": false,
            "in": "query",
            "description": "Filter to only active campaigns",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all campaigns",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all campaigns",
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "operationId": "CampaignsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The campaign has been created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/campaigns/masjid/{masjidId}": {
      "get": {
        "operationId": "CampaignsController_findByMasjid",
        "parameters": [
          {
            "name": "masjidId",
            "required": true,
            "in": "path",
            "description": "Masjid ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of campaigns for the masjid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get campaigns by masjid",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/campaigns/{id}": {
      "get": {
        "operationId": "CampaignsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The campaign",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Campaign not found"
          }
        },
        "summary": "Get a campaign by ID",
        "tags": [
          "campaigns"
        ]
      },
      "patch": {
        "operationId": "CampaignsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The campaign has been updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden resource"
          },
          "404": {
            "description": "Campaign not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a campaign",
        "tags": [
          "campaigns"
        ]
      },
      "delete": {
        "operationId": "CampaignsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The campaign has been deleted"
          },
          "403": {
            "description": "Forbidden resource"
          },
          "404": {
            "description": "Campaign not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/campaigns/{campaignId}/donations": {
      "get": {
        "operationId": "CampaignsController_findDonations",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of donations for the campaign",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DonationResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all donations for a campaign",
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "operationId": "CampaignsController_createDonation",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDonationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The donation has been created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonationResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a donation for a campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
...
  },
  "info": {
    "title": "Masjid Network API",
    "description": "\n# Masjid Network API\n\n## Overview\nThe Masjid Network API provides a comprehensive backend for managing mosques (masjids), fundraising campaigns, donations, and user accounts. This platform connects mosques with their communities and facilitates fundraising for various projects and initiatives.\n\n## Features\n- User authentication and management\n- Masjid/mosque profiles and administration\n- Fundraising campaigns\n- Donation processing\n- Payment handling\n\n## Authentication\nMost endpoints require authentication using JWT Bearer tokens.\n\nTo obtain a token, use the /auth/login endpoint with valid credentials.\nInclude the token in the Authorization header: `Bearer your_token_here`\n\n## Getting Started\n1. Register a user account via /auth/register\n2. Login to obtain a JWT token via /auth/login\n3. Use the token to access protected endpoints\n    ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "user": {
            "type": "object"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CampaignResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the campaign"
          },
          "title": {
            "type": "string",
            "description": "Title of the campaign"
          },
          "description": {
            "type": "string",
            "description": "Description of the campaign"
          },
          "goal": {
            "type": "number",
            "description": "Funding goal amount for the campaign"
          },
          "raised": {
            "type": "number",
            "description": "Current amount raised for the campaign"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "description": "Start date of the campaign"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "End date of the campaign (optional)"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the campaign is active"
          },
          "coverImageUrl": {
            "type": "string",
            "description": "Cover image URL for the campaign"
          },
          "masjidId": {
            "type": "string",
            "description": "ID of the masjid running the campaign"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the campaign was created"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the campaign was last updated"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "goal",
          "raised",
          "startDate",
          "endDate",
          "isActive",
          "coverImageUrl",
          "masjidId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCampaignDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the campaign"
          },
          "description": {
            "type": "string",
            "description": "Description of the campaign"
          },
          "goal": {
            "type": "number",
            "description": "Funding goal amount for the campaign"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "description": "Start date of the campaign"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "End date of the campaign (optional)"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the campaign is active"
          },
          "coverImageUrl": {
            "type": "string",
            "description": "Cover image URL for the campaign"
          },
          "masjidId": {
            "type": "string",
            "description": "ID of the masjid running the campaign"
          }
        },
        "required": [
          "title",
          "description",
          "goal",
          "startDate",
          "endDate",
          "isActive",
          "coverImageUrl",
          "masjidId"
        ]
      },
      "UpdateCampaignDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the campaign"
          },
          "description": {
            "type": "string",
            "description": "Description of the campaign"
          },
          "goal": {
            "type": "number",
            "description": "Funding goal amount for the campaign"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "description": "Start date of the campaign"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "End date of the campaign (optional)"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the campaign is active"
          },
          "coverImageUrl": {
            "type": "string",
            "description": "Cover image URL for the campaign"
          },
          "masjidId": {
            "type": "string",
            "description": "ID of the masjid running the campaign"
          }
        }
      },
      "DonationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the donation"
          },
          "amount": {
            "type": "number",
            "description": "Amount donated"
          },
          "paymentMethod": {
            "type": "string",
            "description": "Payment method used"
          },
          "transactionId": {
            "type": "string",
            "description": "Transaction ID from payment provider"
          },
          "status": {
            "type": "string",
            "description": "Status of the donation",
            "enum": [
              "PENDING",
              "COMPLETED",
              "FAILED",
              "REFUNDED"
            ]
          },
          "isAnonymous": {
            "type": "boolean",
            "description": "Whether the donation is anonymous"
          },
          "message": {
            "type": "string",
            "description": "Optional message from donor"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who made the donation"
          },
          "campaignId": {
            "type": "string",
            "description": "ID of the campaign for this donation"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the donation was created"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the donation was last updated"
          }
        },
        "required": [
          "id",
          "amount",
          "paymentMethod",
          "transactionId",
          "status",
          "isAnonymous",
          "message",
          "userId",
          "campaignId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateDonationDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Amount to donate"
          },
          "paymentMethod": {
            "type": "string",
            "description": "Payment method to use"
          },
          "isAnonymous": {
            "type": "boolean",
            "description": "Whether the donation should be anonymous"
          },
          "message": {
            "type": "string",
            "description": "Optional message to include with donation"
          },
          "campaignId": {
            "type": "string",
            "description": "ID of the campaign to donate to"
          }
        },
        "required": [
          "amount",
          "paymentMethod",
          "isAnonymous",
          "message",
          "campaignId"
        ]
      },
...
    }
  }
}